third_party/WebKit/LayoutTests/fast/css3-text/css3-text-decoration/getComputedStyle/getComputedStyle-text-decoration-shorthand-ordering.html

<!DOCTYPE html>
<html>
    <head>
        <script src="../../../js/resources/js-test-pre.js"></script>
    </head>
    <body>
        <script>
            function testElementStyle(propertyJS, propertyCSS, value)
            {
                shouldBe("e.style.getPropertyCSSValue('" + propertyCSS + "').cssText", "'" + value + "'");
            }

            description("Test to make sure text-decoration longhand values are accepted in every possible ordering in shorthand.")

            var testContainer = document.createElement("div");
            testContainer.contentEditable = true;
            document.body.appendChild(testContainer);

            testContainer.innerHTML = '<div id="test">hello world</div>';

            var e = document.getElementById('test');

            debug("-line -style -color ordering (underline dashed red):");
            e.style.textDecoration = 'underline dashed red';
            testElementStyle("textDecorationLine", "text-decoration-line", "underline");
            testElementStyle("textDecorationStyle", "text-decoration-style", "dashed");
            testElementStyle("textDecorationColor", "text-decoration-color", "red");
            debug("");

            debug("-line -color -style ordering (overline blue dotted):");
            e.style.textDecoration = 'overline blue dotted';
            testElementStyle("textDecorationLine", "text-decoration-line", "overline");
            testElementStyle("textDecorationStyle", "text-decoration-style", "dotted");
            testElementStyle("textDecorationColor", "text-decoration-color", "blue");
            debug("");

            debug("-style -line -color ordering (double line-through underline overline green):");
            e.style.textDecoration = 'double line-through underline overline green';
            testElementStyle("textDecorationLine", "text-decoration-line", "line-through underline overline");
            testElementStyle("textDecorationStyle", "text-decoration-style", "double");
            testElementStyle("textDecorationColor", "text-decoration-color", "green");
            debug("");

            debug("-style -color -line ordering (wavy yellow line-through):");
            e.style.textDecoration = 'wavy yellow line-through';
            testElementStyle("textDecorationLine", "text-decoration-line", "line-through");
            testElementStyle("textDecorationStyle", "text-decoration-style", "wavy");
            testElementStyle("textDecorationColor", "text-decoration-color", "yellow");
            debug("");

            debug("-color -line -style ordering (black underline solid):");
            e.style.textDecoration = 'black underline solid';
            testElementStyle("textDecorationLine", "text-decoration-line", "underline");
            testElementStyle("textDecorationStyle", "text-decoration-style", "solid");
            testElementStyle("textDecorationColor", "text-decoration-color", "black");
            debug("");

            debug("-color style -line -style ordering (navy dashed overline):");
            e.style.textDecoration = 'navy dashed overline';
            testElementStyle("textDecorationLine", "text-decoration-line", "overline");
            testElementStyle("textDecorationStyle", "text-decoration-style", "dashed");
            testElementStyle("textDecorationColor", "text-decoration-color", "navy");
            debug("");

            document.body.removeChild(testContainer);
        </script>
        <script src="../../../js/resources/js-test-post.js"></script>
    </body>
</html>
